fix: reorganizada as pastas e arquivos
