+++ source/srcAdm/funcoesADM.h
@@ -1,0 +1,4 @@
+#ifndef FUNCOES_ADM_H
+#define FUNCOES_ADM_H
+
+#endif
