+++ source/srcAdm/mainADM.c
@@ -1,0 +1,7 @@
+#include <stdio.h>
+
+int main(){
+    printf("Ola");
+
+    return 0;
+}
