+++ source/srcInvestidor/limpaBuffer.c
