+++ source/srcInvestidor/sacar.c
