+++ source/srcInvestidor/venderCripto.c
